+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-- Add SecurityCodeScan tool to build chain for .Net code
+-   Add SecurityCodeScan tool to build chain for .Net code
+-   Extraction report: Group PixelData separately and sort by length
 
 ## [1.11.1] - 2020-08-12